% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_beta.R
\name{align_beta}
\alias{align_beta}
\title{Align Topics}
\usage{
align_beta(src, dest, dist.fn = KLD)
}
\arguments{
\item{src}{a K x V matrix}

\item{dest}{a K x V matrix}

\item{dist.fn}{the distance function that is used to find distance between rows}
}
\value{
\code{align.indices} vector
}
\description{
Align the rows of the \code{dest} matrix in terms of the \code{src} matrix
}
\note{
Created on April 14, 2016
}
