% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLDA.R
\name{CLDA}
\alias{CLDA}
\title{CLDA}
\usage{
CLDA(x, y, linear, type, m, gamma = 0.01)
}
\arguments{
\item{x}{Numeric matrix of features.}

\item{y}{Binary-valued vector of class labels.}

\item{linear}{Boolean: Whether to use LDA (true) or QDA (false).}

\item{type}{Type of compression. One of "compressed", "projected", "FRF", "sub-sampled", or "full".
For QDA, "projected" and "FRF" are not implemented.}

\item{m}{Reduced sample size.}

\item{gamma}{Tuning parameter for numerical stability (see the paper). Defaults to 0.01.}
}
\value{

}
\description{
CLDA
}
