% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_synth_clda_corpus.R
\name{gen_synth_clda_corpus}
\alias{gen_synth_clda_corpus}
\title{Generates a Synthetic c-LDA Corpus}
\usage{
gen_synth_clda_corpus(K, V, J, collection.size, doc.size, alpha.h, gamma.h,
  eta.h)
}
\arguments{
\item{K}{number of topics}

\item{V}{vocabulary size}

\item{J}{number of collections}

\item{collection.size}{number of documents in each collection (a list of J
elements)}

\item{doc.size}{number of words in each document}

\item{alpha.h}{hyperparameter for collection-level Dirichlet sampling}

\item{gamma.h}{hyperparameter for document-level Dirichlet sampling}

\item{eta.h}{hyperparameter for topic Diriclet sampling}
}
\value{
a list of generated corpus and their statistics
}
\description{
Generates documents using the c-LDA generative process based on a set of
predefined values.
}
\details{
Last modified on: March 04, 2016
}
\examples{
## Generates documents with given parameters

J                  <- 2
K                  <- 4
V                  <- 20
alpha.h            <- 2
gamma.h            <- .2
eta.h              <- .25
doc.size           <- 80
collection.size    <- c(40, 40) # number of documents in each collection
ds.name            <- paste("synth-J", J, "-K", K, "-D", D, "-V", V, sep = "")

ds                 <- gen_synth_clda_corpus(K, V, J, collection.size, doc.size, alpha.h, gamma.h, eta.h)

}
\seealso{
Other corpus: \code{\link{gen_synth_clda_corpus_pi}}
}

