% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{H}
\alias{H}
\title{Entropy}
\usage{
H(v)
}
\arguments{
\item{v}{a distribution as a vector}
}
\description{
Entropy
}
\seealso{
Other utils: \code{\link{JSD.matrix}}, \code{\link{JSD}},
  \code{\link{KLD}}, \code{\link{sample_antoniak}},
  \code{\link{sample_dirichlet}},
  \code{\link{sample_multinomial}}
}
