% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{JSD.matrix}
\alias{JSD.matrix}
\title{Jensen-Shannon Divergence (for more than two distributions)}
\usage{
JSD.matrix(w, M)
}
\arguments{
\item{w}{A vector of weights which should sum up to 1}

\item{M}{a matrix with the input distributions as columns}
}
\description{
For more than two distributions, we need a function to compute the Entropy
H(v)
}
\note{
Adapted from the discussion
http://stackoverflow.com/questions/11226627/jensen-shannon-divergence-in-r
}
\seealso{
Other utils: \code{\link{H}}, \code{\link{JSD}},
  \code{\link{KLD}}, \code{\link{sample_antoniak}},
  \code{\link{sample_dirichlet}},
  \code{\link{sample_multinomial}}
}

