% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{JSD}
\alias{JSD}
\title{Jensen-Shannon Divergence (pairwise)}
\usage{
JSD(p, q)
}
\arguments{
\item{p}{a distribution as a vector}

\item{q}{another distribution as a vector}
}
\description{
Jensen-Shannon Divergence (pairwise)
}
\note{
Adapted from the discussion
http://stackoverflow.com/questions/11226627/jensen-shannon-divergence-in-r
}
\seealso{
Other utils: \code{\link{H}}, \code{\link{JSD.matrix}},
  \code{\link{KLD}}, \code{\link{sample_antoniak}},
  \code{\link{sample_dirichlet}},
  \code{\link{sample_multinomial}}
}
