% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{KLD}
\alias{KLD}
\title{Kullback-Leibler Divergence}
\usage{
KLD(p, q)
}
\arguments{
\item{p}{a distribution as a vector}

\item{q}{another distribution as a vector}
}
\description{
Kullback-Leibler Divergence
}
\note{
Adapted from the discussion
http://enterotype.embl.de/enterotypes.html
}
\seealso{
Other utils: \code{\link{H}}, \code{\link{JSD.matrix}},
  \code{\link{JSD}}, \code{\link{sample_antoniak}},
  \code{\link{sample_dirichlet}},
  \code{\link{sample_multinomial}}
}
