% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misclass_CLDA.R
\name{misclass.CLDA}
\alias{misclass.CLDA}
\title{Compute misclassification error rates}
\usage{
misclass.CLDA(preds, y)
}
\arguments{
\item{preds}{Vector of predictions}

\item{y}{Vector of true values}
}
\value{

}
\description{
Compute misclassification error rates
}
