% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{news}
\alias{news}
\title{16 Newsgroups Dataset}
\format{\code{vocab} a vector of unique words in the corpus vocabulary.

\code{docs} a list of documents in the corpus. Each item (represents a
 document) is a matrix (2 X U) of word frequencies, where U represents the
 number of unique words in a document. Each column in the matrix represents
 a unique word in a document and contains
 \itemize{
   \item vocabulary-id. the index of the word in the vocabulary (starts with 0)
   \item frequency. the relative frequency of the word in the document
 }

\code{docs.metadata} a matrix of document (article) metadata, where each
row represents a document with
 \itemize{
   \item category. the category assigned to the article
   \item name. the name of the news article from the 20Newsgroups dataset
   \item doclength. the number of words in the article
   \item collection. the collection name of each article
 }

\code{cids} a vector of document collection IDs

\code{class.labels} a vector of categories (classes) in the corpus

\code{collection.labels} a vector of collections in the corpus

\code{ds.name} the corpus name (string)

\code{num.docs} the number of documents in the corpus

\code{V} the vocabulary size}
\source{
Articles are downloaded via
\href{http://scikit-learn.org/}{scikit-learn}
}
\usage{
\code{data("news")}
}
\description{
It's a subset of the
\href{http://qwone.com/~jason/20Newsgroups}{20Newsgroups} dataset. This
corpus consists of 10,764 news articles and 9,208 unique words.
}
\note{
Created on July 26, 2015
}
\author{
Clint P. George
}
\seealso{
Other datasets: \code{\link{nips}}, \code{\link{yelp}}
}

