% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{nips}
\alias{nips}
\title{NIPS 00-18 Dataset}
\format{\code{vocab} a vector of unique words in the corpus vocabulary.

\code{docs} a list of documents in the corpus. Each item (represents a
 document) is a matrix (2 X U) of word frequencies, where U represents the
 number of unique words in a document. Each column in the matrix represents
 a unique word in a document and contains
 \itemize{
   \item vocabulary-id. the index of the word in the vocabulary (starts with 0)
   \item frequency. the relative frequency of the word in the document
 }

\code{docs.metadata} a matrix of document (article) metadata, where each
row represents a document with
 \itemize{
   \item doc.id. a unique article id
   \item file.name. the name of the article
   \item row.word.count. the number of words in the article
   \item collection. the collection name of each article
 }

\code{cids} a vector of document collection IDs

\code{class.labels} a vector of categories (classes) in the corpus

\code{collection.labels} a vector of collections in the corpus

\code{ds.name} the corpus name (string)

\code{num.docs} the number of documents in the corpus

\code{V} the vocabulary size}
\source{
Articles are downloaded from the
 \href{http://robotics.stanford.edu/~gal/data.html}{link}
}
\usage{
\code{data("nips")}
}
\description{
It's a subset of the NIPS dataset (Chechik 2007). It consists of papers
published in proceedings 00 to 18 of the  Neural Information Processing
Systems (NIPS) conference (i.e. years from 1988 to 2005). This corpus
consists of 2,741 news articles and 9,156 unique words.
}
\note{
Created on July 26, 2015
}
\seealso{
Other datasets: \code{\link{news}}, \code{\link{yelp}}
}
\author{
Clint P. George
}
