% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_CLDA.R
\name{predict.CLDA}
\alias{predict.CLDA}
\title{predict.CLDA
Make predictions on a new set of data from a CLDA object.}
\usage{
\method{predict}{CLDA}(mod, x)
}
\arguments{
\item{mod}{CLDA object}

\item{x}{New data}
}
\value{

}
\description{
predict.CLDA
Make predictions on a new set of data from a CLDA object.
}
