% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rademacher.R
\name{Rademacher}
\alias{Rademacher}
\title{Create a Rademacher matrix}
\usage{
Rademacher(nrow, ncol, s)
}
\arguments{
\item{nrow}{Number of rows.}

\item{ncol}{Number of columns.}

\item{s}{Sparsity parameter in (0, 1). Probability of an entry being zero is 1-s.}
}
\value{

}
\description{
Create a Rademacher matrix
}
