% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clda.R
\docType{package}
\name{clda}
\alias{clda}
\alias{clda-package}
\alias{package-clda}
\title{Algorithms for the Compound Latent Dirichlet Allocation Model}
\description{
This R package implements the following inference algorithms
for the compound latent Dirichlet allocation (CLDA) model.
\itemize{
	\item Auxiliary variable update within Gibbs sampler (AGS)
 \item Metropolis adjusted Langevin algorithm within Gibbs sampler (MGS)
 \item Variational Expectation Maximization (VEM)
}
}
\author{
Clint P. George, Wei Xia, and George Michailidis
}

